vehicle and steeringbehaviours have some behaviour

+++ AIproject/AIproject/AIproject/BaseGameEntity.cs
@@ -8,7 +8,7 @@ namespace AIproject
 {
     abstract class BaseGameEntity
     {
-        public abstract void Update(double time_elapsed);
+        public abstract void Update(float time_elapsed);
         public abstract void Render();
         public int ID { get; set; }
         public Vector2 pos { get; set; }
